requirements.txt
fastapi~=0.109.0
sqlalchemy~=2.0.25
alembic~=1.13.1
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
pydantic~=2.5.3
uvicorn[standard]~=0.26.0
python-multipart~=0.0.6
